% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solids.R
\name{centricCuboctahedron}
\alias{centricCuboctahedron}
\title{Centric cuboctahedron}
\usage{
centricCuboctahedron()
}
\value{
A numeric matrix with 13 rows and 3 columns.
}
\description{
A cuboctahedron (12 vertices), with a point added at its center.
}
