% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{delaunay}
\alias{delaunay}
\title{Delaunay triangulation}
\usage{
delaunay(points, atinfinity = FALSE, degenerate = FALSE)
}
\arguments{
\item{points}{the points given as a matrix, one point per row}

\item{atinfinity}{Boolean, whether to include a point at infinity}

\item{degenerate}{Boolean, whether to include degenerate tiles}
}
\value{
The Delaunay tessellation with many details, in a list.
}
\description{
Delaunay triangulation (or tesellation) of a set of points.
}
\examples{
library(tesselation)
points <- rbind(
 c(0.5,0.5,0.5),
 c(0,0,0),
 c(0,0,1),
 c(0,1,0),
 c(0,1,1),
 c(1,0,0),
 c(1,0,1),
 c(1,1,0),
 c(1,1,1)
)
delaunay(points)
}
\seealso{
\code{\link{getDelaunaySimplicies}}
}
