% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiEdges.R
\name{IEdge}
\alias{IEdge}
\title{R6 class representing an infinite edge of a Voronoï cell}
\description{
An infinite edge is given by a vertex, its origin,
  and a vector, its direction.
}
\examples{

## ------------------------------------------------
## Method `IEdge$new`
## ------------------------------------------------

iedge <- IEdge$new(c(1, 1, 1), c(1, 2, 3))
edge
edge$O
edge$O <- c(1, 0, 0)
edge

## ------------------------------------------------
## Method `IEdge$print`
## ------------------------------------------------

IEdge$new(c(2, 0, 0), c(3, -1, 4))
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{O}}{get or set the vertex \code{O}}

\item{\code{direction}}{get or set the vector \code{direction}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IEdge$new()}}
\item \href{#method-print}{\code{IEdge$print()}}
\item \href{#method-clone}{\code{IEdge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{IEdge} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IEdge$new(O, direction)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{O}}{the vertex \code{O} (origin)}

\item{\code{direction}}{the vector \code{direction}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{IEdge} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{iedge <- IEdge$new(c(1, 1, 1), c(1, 2, 3))
edge
edge$O
edge$O <- c(1, 0, 0)
edge
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Show instance of an \code{IEdge} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IEdge$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{IEdge$new(c(2, 0, 0), c(3, -1, 4))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IEdge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
