% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{isBoundedCell}
\alias{isBoundedCell}
\title{Is this cell bounded?}
\usage{
isBoundedCell(cell)
}
\arguments{
\item{cell}{a Voronoï cell}
}
\value{
A Boolean value, whether the cell is bounded.
}
\description{
Check whether a Voronoï cell is bounded, i.e. contains only
  finite edges.
}
