% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plotBoundedCell}
\alias{plotBoundedCell}
\title{Plot a bounded Voronoï cell}
\usage{
plotBoundedCell(
  cell,
  edgesAsTubes = FALSE,
  tubeRadius,
  tubeColor,
  facetsColor = NULL,
  alpha = 1
)
}
\arguments{
\item{cell}{a bounded Voronoï cell}

\item{edgesAsTubes}{Boolean, whether to plot edges as tubes or as lines}

\item{tubeRadius}{radius of the tubes if \code{edgesAsTubes = TRUE}}

\item{tubeColor}{color of the tubes if \code{edgesAsTubes = TRUE}}

\item{facetsColor}{color of the facets; \code{NULL} for no color}

\item{alpha}{opacity of the facets, between 0 and 1}
}
\value{
No value, just plots the cell.
}
\description{
Plot a bounded Voronoï cell with \emph{rgl}.
}
\examples{
library(tessellation)
d <- delaunay(centricCuboctahedron())
v <- voronoi(d)
cell13 <- v[[13]]
isBoundedCell(cell13) # TRUE
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotBoundedCell(
  cell13, edgesAsTubes = TRUE, tubeRadius = 0.1, tubeColor = "yellow",
  facetsColor = "navy", alpha = 0.3
)
}
