% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{plotDelaunay3D}
\alias{plotDelaunay3D}
\title{Plot 3D Delaunay tessellation}
\usage{
plotDelaunay3D(
  tessellation,
  color = TRUE,
  hue = "random",
  luminosity = "light",
  alpha = 0.3,
  exteriorEdgesAsTubes = FALSE,
  tubeRadius,
  tubeColor
)
}
\arguments{
\item{tessellation}{the output of \code{\link{delaunay}}}

\item{color}{Boolean, whether to use colors}

\item{hue, luminosity}{if \code{color = TRUE}, these arguments are passed to
\code{\link[randomcoloR]{randomColor}}}

\item{alpha}{opacity, number between 0 and 1}

\item{exteriorEdgesAsTubes}{Boolean, whether to plot the exterior edges
as tubes}

\item{tubeRadius}{if \code{exteriorEdgesAsTubes = TRUE}, the radius of
the tubes}

\item{tubeColor}{if \code{exteriorEdgesAsTubes = TRUE}, the color of
the tubes}
}
\value{
No value, just renders a 3D plot.
}
\description{
Plot a 3D Delaunay tessellation with \strong{rgl}.
}
\note{
The option \code{exteriorEdgesAsTubes} is experimental. I found an
  example for which it renders some tubes at the interior of the convex
  hull. However I do not remove it from the package since it seems to
  work in many cases.
}
\examples{
library(tessellation)
pts <- rbind(
  c(-5, -5,  16),
  c(-5,  8,   3),
  c(4,  -1,   3),
  c(4,  -5,   7),
  c(4,  -1, -10),
  c(4,  -5, -10),
  c(-5,  8, -10),
  c(-5, -5, -10)
)
tess <- delaunay(pts)
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotDelaunay3D(tess)
open3d(windowRect = c(50, 50, 562, 562))
plotDelaunay3D(
  tess, exteriorEdgesAsTubes = TRUE, tubeRadius = 0.3, tubeColor = "yellow"
)
}
