% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plotVoronoiDiagram}
\alias{plotVoronoiDiagram}
\title{Plot Voronoî diagram}
\usage{
plotVoronoiDiagram(
  v,
  color = TRUE,
  hue = "random",
  luminosity = "light",
  alpha = 1,
  ...
)
}
\arguments{
\item{v}{an output of \code{\link{voronoi}}}

\item{color}{Boolean, whether to use colors}

\item{hue, luminosity}{if \code{color = TRUE}, these arguments are passed to
\code{\link[randomcoloR]{randomColor}}}

\item{alpha}{opacity, a number between 0 and 1
(used when \code{color = TRUE})}

\item{...}{arguments passed to \code{\link{plotBoundedCell2D}} or
\code{\link{plotBoundedCell3D}}}
}
\value{
No returned value.
}
\description{
Plot all the bounded cells of a 2D or 3D Voronoï tessellation.
}
\examples{
library(tessellation)
# 2D example: sunflower
pts <- sunflower(150L)
d <- delaunay(pts)
v <- voronoi(d)
opar <- par(mar = c(0, 0, 0, 0))
plot(pts, type = "n", xlab = NA, ylab = NA, asp = 1, axes = FALSE)
plotVoronoiDiagram(
  v, luminosity = "dark"
)
par(opar)

}
