% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sunflower}
\alias{sunflower}
\title{Sunflower}
\usage{
sunflower(nmax)
}
\arguments{
\item{nmax}{maximum value of the index (try it on the example)}
}
\value{
A numeric matrix with \code{nmax} rows and two columns.
}
\description{
Returns a cloud of points looking like a sunflower.
}
\examples{
pts <- sunflower(150L)
opar <- par(mar = c(0, 0, 0, 0))
plot(pts, pch = 19, asp = 1, axes = FALSE, xlab = NA, ylab = NA)
}
