% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{surface}
\alias{surface}
\title{Tessellation surface}
\usage{
surface(tessellation)
}
\arguments{
\item{tessellation}{output of \code{\link{delaunay}}}
}
\value{
A number, the exterior surface of the Delaunay tessellation.
}
\description{
Exterior surface of the Delaunay tessellation.
}
\seealso{
\code{\link{volume}}
}
