% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\docType{import}
\name{tessellation-imports}
\alias{tessellation-imports}
\alias{values}
\alias{keys}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages.
  Follow the links to their documentation:
  \code{\link[hash:values]{values}},
  \code{\link[hash:keys]{keys}}.
}
