% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{volume}
\alias{volume}
\title{Tessellation volume}
\usage{
volume(tessellation)
}
\arguments{
\item{tessellation}{output of \code{\link{delaunay}}}
}
\value{
A number, the volume of the Delaunay tessellation (area in 2D).
}
\description{
The volume of the Delaunay tessellation, that is, the volume of
  the convex hull of the sites.
}
\seealso{
\code{\link{surface}}
}
