% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Voronoï tessellation}
\usage{
voronoi(tessellation)
}
\arguments{
\item{tessellation}{output of \code{\link{delaunay}}}
}
\value{
A list representing the Voronoï tessellation.
}
\description{
Voronoï tessellation from Delaunay tessellation; this is a list
  of pairs made of a site (a vertex) and a list of edges.
}
\examples{
library(tessellation)
d <- delaunay(centricCuboctahedron())
voronoi(d)
}
