% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiEdges.R
\name{Edge}
\alias{Edge}
\title{R6 class representing an edge of a Voronoï cell}
\description{
An edge is given by two vertices,
  named \code{A} and \code{B}.
}
\examples{

## ------------------------------------------------
## Method `Edge$new`
## ------------------------------------------------

edge <- Edge$new(c(1, 1, 1), c(1, 2, 3))
edge
edge$A
edge$A <- c(1, 0, 0)
edge

## ------------------------------------------------
## Method `Edge$print`
## ------------------------------------------------

Edge$new(c(2, 0, 0), c(3, -1, 4))
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{A}}{get or set the vertex \code{A}}

\item{\code{B}}{get or set the vertex \code{B}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Edge$new()}}
\item \href{#method-print}{\code{Edge$print()}}
\item \href{#method-plot}{\code{Edge$plot()}}
\item \href{#method-clone}{\code{Edge$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Edge} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$new(A, B)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{A}}{the vertex \code{A}}

\item{\code{B}}{the vertex \code{B}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Edge} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{edge <- Edge$new(c(1, 1, 1), c(1, 2, 3))
edge
edge$A
edge$A <- c(1, 0, 0)
edge
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Show instance of an \code{Edge} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Edge$new(c(2, 0, 0), c(3, -1, 4))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Edge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
