% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{plotDelaunay3D}
\alias{plotDelaunay3D}
\title{Plot 3D Delaunay tessellation}
\usage{
plotDelaunay3D(
  tessellation,
  color = TRUE,
  hue = "random",
  luminosity = "light",
  alpha = 0.3
)
}
\arguments{
\item{tessellation}{the output of \code{\link{delaunay}}}

\item{color}{Boolean, whether to use colors}

\item{hue, luminosity}{if \code{color = TRUE}, these arguments are passed to
\code{\link[randomcoloR]{randomColor}}}

\item{alpha}{opacity, number between 0 and 1}
}
\value{
No value, just renders a 3D plot.
}
\description{
Plot a 3D Delaunay tessellation with \strong{rgl}.
}
\examples{
library(tessellation)
pts <- rbind(
  c(-5, -5,  16),
  c(-5,  8,   3),
  c(4,  -1,   3),
  c(4,  -5,   7),
  c(4,  -1, -10),
  c(4,  -5, -10),
  c(-5,  8, -10),
  c(-5, -5, -10)
)
tess <- delaunay(pts)
options(rgl.useNULL = TRUE)
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotDelaunay3D(tess)
}
