% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{plotDelaunay3D}
\alias{plotDelaunay3D}
\title{Plot 3D Delaunay tessellation}
\usage{
plotDelaunay3D(tesselation, color = TRUE, alpha = 0.3)
}
\arguments{
\item{tesselation}{the output of \code{\link{delaunay}}}

\item{color}{Boolean, whether to use colors}

\item{alpha}{opacity, between 0 and 1}
}
\value{
No value, just renders a 3D plot.
}
\description{
Plot a 3D Delaunay tessellation with \emph{rgl}.
}
\examples{
library(tessellation)
pts <- rbind(
  c(-5, -5,  16),
  c(-5,  8,   3),
  c(4,  -1,   3),
  c(4,  -5,   7),
  c(4,  -1, -10),
  c(4,  -5, -10),
  c(-5,  8, -10),
  c(-5, -5, -10)
)

tess <- delaunay(pts)
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotDelaunay3D(tess)
}
