% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{plotVoronoiDiagram}
\alias{plotVoronoiDiagram}
\title{Plot Voronoï diagram}
\usage{
plotVoronoiDiagram(
  v,
  colors = "random",
  hue = "random",
  luminosity = "light",
  alpha = 1,
  ...
)
}
\arguments{
\item{v}{an output of \code{\link{voronoi}}}

\item{colors}{this can be \code{"random"} to use random colors for the cells,
\code{NA} for no colors, or a vector of colors; the length of this vector
of colors must match the number of bounded cells, that you can get by
typing \code{length(Filter(isBoundedCell, v))}}

\item{hue, luminosity}{if \code{colors = "random"}, these arguments are passed
to \code{\link[randomcoloR]{randomColor}}}

\item{alpha}{opacity, a number between 0 and 1
(used when \code{colors} is not \code{NA})}

\item{...}{arguments passed to \code{\link{plotBoundedCell2D}} or
\code{\link{plotBoundedCell3D}}}
}
\value{
No returned value.
}
\description{
Plot all the bounded cells of a 2D or 3D Voronoï tessellation.
}
\note{
Sometimes, it is necessary to set the option \code{degenerate=TRUE}
  in the \code{\link{delaunay}} function in order to get a correct
  Voronoï diagram with the \code{plotVoronoiDiagram} function (I don't know
  why).
}
\examples{
library(tessellation)
# 2D example: sunflower surrounded by a circle
angles <- seq(0, 2*pi, length.out=91)[-1]
circle <- 25000 * cbind(cos(angles), sin(angles))
pts <- rbind(sunflower(50L, 150L), circle)
d <- delaunay(pts)
v <- voronoi(d)
opar <- par(mar = c(0, 0, 0, 0))
plot(pts, type = "n", xlab = NA, ylab = NA, asp = 1, axes = FALSE)
plotVoronoiDiagram(
  v, luminosity = "dark"
)
par(opar)

# 3D example: tetrahdron surrounded by three circles
tetrahedron <-
  rbind(
    c(2*sqrt(2)/3, 0, -1/3),
    c(-sqrt(2)/3, sqrt(2/3), -1/3),
    c(-sqrt(2)/3, -sqrt(2/3), -1/3),
    c(0, 0, 1)
  )
angles <- seq(0, 2*pi, length.out = 91)[-1]
R <- 2.5
circle1 <- t(vapply(angles, function(a) R*c(cos(a), sin(a), 0), numeric(3L)))
circle2 <- t(vapply(angles, function(a) R*c(cos(a), 0, sin(a)), numeric(3L)))
circle3 <- t(vapply(angles, function(a) R*c(0, cos(a), sin(a)), numeric(3L)))
circles <- rbind(circle1, circle2, circle3)
pts <- rbind(tetrahedron, circles)
d <- delaunay(pts, degenerate = TRUE)
v <- voronoi(d)
library(rgl)
open3d(windowRect = c(50, 50, 562, 562))
plotVoronoiDiagram(v, luminosity = "bright")
}
