% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sunflower}
\alias{sunflower}
\title{Sunflower}
\usage{
sunflower(nmin, nmax)
}
\arguments{
\item{nmin, nmax}{range of the index (try on the example)}
}
\value{
A numeric matrix of 2D points (rows).
}
\description{
Returns a cloud of points looking like a sunflower.
}
\examples{
pts <- sunflower(50L, 150L)
opar <- par(mar = c(0, 0, 0, 0))
plot(pts, pch = 19, asp = 1, axes = FALSE, xlab = NA, ylab = NA)
}
